lint: clear every clippy backlog, real this time, and gate all of it
components/'s clippy CI job has been advisory since #112, with a
comment promising `-D warnings` "the day" the backlog it cleared to 22
was gone. It never was flipped, and the backlog grew back to 125+ as
~100 more components landed after #112 with nobody re-clearing their
warnings. host, lattice, reconciler, cli, xtask and comp-metadata were
never checked at all — not gated, not advisory, nothing (104 warnings
in reconciler alone).

Cleared all of it — 0 warnings across all ~200 components/ crates and
all 7 native-workspace crates — auto-fixed where clippy could, by hand
where it couldn't:

  - `while_let_loop` rewrites (a paging loop written as `loop { let-else
    ... }` instead of `while let`), in ~8 crates
  - `type_complexity`: named type aliases for tuples/maps that appeared
    once in a signature and nowhere else, never a new abstraction
  - `too_many_arguments`: `#[allow]` with a one-line reason where the
    args are genuinely distinct (a generic search's callbacks, a
    booking's fields) rather than a struct waiting to be extracted
  - `manual_checked_ops`: four hand-rolled `if total == 0 { 0 } else
    { n / total }` divisions (three domains' own scoring, plus
    comp-checks' — found only by running this, since it isn't
    clippy-gated either) become `.checked_div(total).unwrap_or(0)`
  - a few genuine small findings the lints pointed at: a zombie
    nats-server left behind by a test's panic path (no `.wait()` on the
    abort branch), and two truly orphaned doc comments (one documented
    a function defined in a different file entirely, left behind from
    before an import replaced a duplicated copy)

Every touched crate's own tests still pass (reconciler: 190 lib tests
plus every touched bin's test suite, components: every touched crate
individually, cli: 70, lattice: 9, comp-metadata: 1) — these are
lint-driven rewrites verified equivalent by the type checker and the
test suite, not a behavior change.

CI: components/'s job is finally `-D warnings`, for real, with the
history of why it wasn't recorded in the comment this time so the next
regression is not a surprise. A new `lint-native` job gates the six
workspaces that had no clippy coverage at all.

+++ components/luhn-checksum/src/lib.rs
@@ -47,7 +47,7 @@ pub fn checksum_digit(digits: &str) -> Option<u8> {
     // Once a digit is appended, every existing digit's doubling flips (the
     // rightmost existing digit is no longer the rightmost overall).
     let sum = folded_sum(digits, true)?;
-    Some(((10 - (sum % 10) as u8) % 10) as u8)
+    Some((10 - (sum % 10) as u8) % 10)
 }
 
 /// `digits` with its check digit appended, or `None` on the same bad input as
